[Bug 68] Account for linker expansion of PC-relative branches
Jump-table relaxation ignores the possibility of further linker
relaxation of some PC-relative branches.  This leads to a failure
to expand some byte-sized jump-tables to half-word and link-time
relocation overflows.  It affects jump-tables having a lot of
local PC-relative branches, particularly branches that are not
relaxed by the assembler like branch-on-immediate-compare.
This is a fall-out of a previous attempt to decouple the
minimize-relocs step from jump-table optimization.

gas/
	* config/tc-nanomips.c (frag_subtype_to_relax_sop): Move
	definition up in source file.
	(frag_subtype_to_relax_sink): New.
	(record_insn_sops): Handle assembler-relaxable instructions.
	(append_insn): Include all PC-relative instructions to estimate
	maximum linker expansion range when minimize-relocs is disabled.
	(mark_if_invariable_branch): Remove double-counting on external
	refernces already counted in append_insn.
	* gas/testsuite/gas/relax-jumptable-insn2.s: New test source.
	* gas/testsuite/gas/relax-jumptable-insn2.d: New test.
	* gas/testsuite/gas/nanomips.exp: Run new test.

Change-Id: Iea5dda6922e695606244613fce946ae5d56e83ae