[nanoMIPS] Just symbols script parsing
Added a feature to pass a textual script after --just-symbols to
collect those symbols and potentially invoke a duplicate symbol
definition error. Symbols should be in this format:

{symbol_name} = {addresss};

Where {symbol_name} is a string and address is an absolute value
written decimal or hexadecimal.

This feature is controlled by the -z nobfd-compat-muldefs option
which is off by default.

For this feature, a new type of script is added in the yyscript.y, the
just symbols script, whose symbols are parsed and added to the symbol
table as CONSTANT symbols, these symbols can trigger a multiple symbol
definition error. The default behaviour
(without -z nobfd-compat-muldefs) just overwrites symbols with the
same name with the ones defined in the script.

Change-Id: I17baa50cfaac45a635559741691949c0abbc567d